+++ exonum/src/blockchain/service.rs
@@ -183,6 +183,10 @@ pub trait Service: Send + Sync + 'static {
         Value::Null
     }
 
+    /// Handles block execution. This handler is invoked for each service after execution
+    /// of all transactions in the mempool but before `handle_commit` handler.
+    fn handle_execute(&self, fork: &mut Fork) {}
+
     /// Handles block commit. This handler is invoked for each service after commit of the block.
     /// For example, a service can create one or more transactions if a specific condition
     /// has occurred.